feat: 입찰가 제어 함수를 훅으로 분리
